requirements.txt
dataclasses==0.6
Django==2.2.1
django-filter==2.1.0
django-split-settings==0.3.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djongo==1.2.32
Markdown==3.1.1
mongoengine==0.17.0
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
pymongo==3.8.0
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.2.4
